requirements.txt
aiofiles==25.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
emoji==2.15.0
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
magic-filter==1.0.12
multidict==6.7.1
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
setuptools==82.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.47
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.6.3
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
